build fiber for Linux ok

+++ lib_fiber/c/src/event.c
@@ -179,7 +179,7 @@ int event_add_read(EVENT *ev, FILE_EVENT *fe, event_proc *proc)
 	}
 
 	if (fe->fd >= (socket_t) ev->setsize) {
-		msg_error("fd: %d >= setsize: %d", fe->fd, ev->setsize);
+		msg_error("fd: %d >= setsize: %d", fe->fd, (int) ev->setsize);
 		acl_fiber_set_error(ERANGE);
 		return 0;
 	}
@@ -218,7 +218,7 @@ int event_add_write(EVENT *ev, FILE_EVENT *fe, event_proc *proc)
 	}
 
 	if (fe->fd >= (socket_t) ev->setsize) {
-		msg_error("fd: %d >= setsize: %d", fe->fd, ev->setsize);
+		msg_error("fd: %d >= setsize: %d", fe->fd, (int) ev->setsize);
 		acl_fiber_set_error(ERANGE);
 		return 0;
 	}
